refactor(backend-cuda): split — extract BackendInt8KvOps into cuda/int8_kv.rs (#8 Phase 1)
`crates/ferrum-kernels/src/backend/cuda.rs` was 5000 lines in one file —
the largest single-file maintenance burden in the codebase. Every
edit to a Backend trait method had to find its home in 5000 lines.
Audit item #8 plan: convert `cuda.rs` to `cuda/mod.rs` (a directory
module) and extract per-supertrait `impl` blocks into peer files.

Phase 1 — smallest, newest, least entangled supertrait first:
- `BackendKvDtype<KvInt8>` marker
- `OptionalCudaInt8` + `OptionalCudaScalesF16` lazy-allocated buffers
- `impl BackendInt8KvOps for CudaBackend` (alloc_paged_int8_layer +
  int8_kv_append_paged + int8_paged_decode_attention)
- `KvCacheQuant<CudaBackend, KvInt8>::new_paged_cuda(...)` constructor

Total ~242 lines moved into `cuda/int8_kv.rs`; `cuda/mod.rs` re-exports
`OptionalCudaInt8` / `OptionalCudaScalesF16` so existing crate-level
consumers keep working unchanged. Pure file move + adding `pub mod
int8_kv;` to `cuda/mod.rs` — zero behaviour change.

Mac local: `cargo check -p ferrum-kernels` ✅, `cargo check --workspace
--features metal` ✅.
